Replace signal error fix + weekly report
